Added Facebook friends and user form components and HTML
This commit is to add the Facebook friends component. This is an angular
component that will eventually go to the Facebook API and populate a
widget that lists friends for the inputted facebook user.  There is also
a user form component. This is an Angular component that uses Reactive
Forms to display a user registration form.